refactor(app, store, command): the line is every identifier, not every string identifier
Four maps keyed by an identifier were left untyped, and the reason given
for leaving them did not cover any of them. The stated line was that
configuration ids and the status document's wire shape stay strings; none
of these is either.

Two are keyed by a binding id, which is an int64 rather than a string, so
an unstated rule had crept in — string-keyed maps only. It is not a rule:
a binding id passed where a delivery id belongs is the same mistake and
the same silent empty result. The snapshot's queue counts and the
controller's binding index carry the type now, and four int64 conversions
went with them.

One is a tier id, in a map beside an allocator that already keys by
assignment.TierID. One is a lease id on a cache lane, in the same
Snapshot whose two other maps this branch typed, and in the same shape as
the Docker fields it converted.

Three identity conversions survived on a field this branch had just
typed, three lines below a map it had converted — exactly the noise it
exists to remove. And two new imports landed inside the standard library
run rather than the project group; gofmt accepts both, so nothing said
so.

One map keyed by a workload key stays a string, and the reason is worth
recording rather than leaving to a reader: assignment.WorkloadAssignment
holds that key as a bare string itself, so typing the map would put a
conversion at the map instead of at the field. The field is where it
belongs, and that is a different change.

+++ internal/app/bindings.go
@@ -114,7 +114,7 @@ func (s *Controller) buildBindings(ctx context.Context, cfg *config.Config, envi
 				return err
 			}
 			s.bindings = append(s.bindings, b)
-			s.byBinding[int64(bindingID)] = b
+			s.byBinding[bindingID] = b
 			claimed = append(claimed, bindingID)
 		}
 	}
